requirements.txt
Cerberus==0.4.0
Events==0.2.0
Fabric==1.8.0
Flask==0.10.1
Flask-Bootstrap==3.0.0.1
Flask-PyMongo==0.3.0
Jinja2==2.7
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
arrow==0.4.2
ecdsa==0.10
itsdangerous==0.22
jedi==0.7.0
paramiko==1.12.0
py==1.4.18
pyOpenSSL==0.13.1
pycrypto==2.6.1
pymongo==2.6.2
pytest==2.4.2
python-dateutil==2.2
requests==2.0.1
simplejson==3.3.0
six==1.4.1
wsgiref==0.1.2
git+git://github.com/ufo22940268/eve.git@master
epc
xlwt
